Story/VSPC-113 (#140)
* [VSPC-113] Moving spaceStatus and showUnpLinks to Space model

* [VSPC-113] adding access annotation for setter

* [VSPC-113] Adding accesstype field for showUnpubLinks

* [VSPC-113] changing accesstype field to property
