+++ vspace/src/main/java/edu/asu/diging/vspace/core/model/IVSpaceElement.java
@@ -2,8 +2,6 @@ package edu.asu.diging.vspace.core.model;
 
 import java.time.OffsetDateTime;
 
-import edu.asu.diging.vspace.core.model.impl.SpaceStatus;
-
 public interface IVSpaceElement {
 
     String getId();
@@ -33,13 +31,5 @@ public interface IVSpaceElement {
     OffsetDateTime getModificationDate();
 
     void setModificationDate(OffsetDateTime modificationDate);
-
-    SpaceStatus getSpaceStatus();
-
-    void setSpaceStatus(SpaceStatus status);
-
-    boolean isShowUnpublishedLinks();
-
-    void setShowUnpublishedLinks(Boolean showUnpublishedLinks);
 
 }
